oop-data-ingestion-tests-aws/pyproject.toml
[tool.poetry]
name = "oop-data-ingestion-tests-aws"
version = "0.1.0"
description = ""
authors = ["Willian Barbosa <devwdougherty@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.1"
schedule = "^1.1.0"
ratelimit = "^2.2.1"
backoff = "^2.2.1"
boto3 = "^1.26.19"
python-dotenv = "^0.21.0"
pynamodb = "^5.3.3"

[tool.poetry.dev-dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
